data/files/PDFmerger_DataStructure/scrapping task perform by fsds students/ineuron scrapping/shivam/requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.1
Flask-Cors==3.0.10
pymongo==3.12.0
dnspython==1.16.0
python-dotenv==0.20.0
requests==2.27.1
gunicorn